breaks affix to its own component
